DEV: Added the data/ folder in tomography to .gitignore
The data folder was created to hold the various types of tomo data that is being used in the example notebooks. This way there is a single data src instead of populating each folder with copies of the source data for various examples. Basically, in an attempt to save space...

This will hopefully also solve the issue with .yaml file metadata dictionaries showing as options for inclusion in version control (GIT)